Moved plot_by_cycle logic to cpp (cherry picked from commit 5179ad3)
